code/requirements.txt
aiofiles==22.1.0
celery==5.2.7
fastapi[all]==0.89.1
requests==2.28.2
uvicorn[standard]==0.20.0
redis==4.4.2
pymongo==4.3.3
pillow==9.4.0
